Move autogen-graphql code into its own directory.
